+++ src/gax-internal/src/grpc/grpc_rust/bidi.rs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub use super::receive::ReceiveTask;
 use bytes::{Buf, Bytes};
-use grpc::StatusCodeError;
-use grpc::client::{RecvStream, ResponseStreamItem};
-use grpc::core::{RecvMessage, SendMessage, Trailers};
-use prost::Message;
+use grpc::core::SendMessage;
 
 /// A [`SendMessage`] adapter that encodes Prost protobuf messages for `grpc-rust`.
 pub struct GrpcRustSend<T>(pub T);
@@ -30,170 +28,11 @@ where
     }
 }
 
-/// A [`RecvMessage`] adapter that decodes raw byte payloads into Prost protobuf messages.
-struct GrpcRustRecv<T>(Option<T>);
-
-impl<T> Default for GrpcRustRecv<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<T> GrpcRustRecv<T> {
-    /// Creates a new unpopulated [`GrpcRustRecv`] receiver.
-    const fn new() -> Self {
-        Self(None)
-    }
-
-    /// Takes the decoded message.
-    fn take(&mut self) -> tonic::Result<T> {
-        self.0.take().ok_or_else(|| {
-            tonic::Status::internal("grpc-rust response message missing or already taken")
-        })
-    }
-}
-
-impl<T> RecvMessage for GrpcRustRecv<T>
-where
-    T: prost::Message + Default,
-{
-    fn decode(&mut self, data: &mut dyn Buf) -> std::result::Result<(), String> {
-        self.0 = Some(T::decode(data).map_err(|e| e.to_string())?);
-        Ok(())
-    }
-}
-
-/// Converts gRPC response [`Trailers`] into a [`tonic::Status`], preserving status code, message, metadata, and `grpc-status-details-bin`.
-pub(super) fn trailers_to_tonic_status(trailers: Trailers) -> Option<tonic::Status> {
-    let status = trailers.status().as_ref().err()?;
-    let metadata: tonic::metadata::MetadataMap = trailers.metadata().clone().into();
-    let details = metadata
-        .get_bin("grpc-status-details-bin")
-        .and_then(|value| value.to_bytes().ok())
-        .unwrap_or_default();
-    Some(tonic::Status::with_details_and_metadata(
-        grpc_rust_error_to_tonic_code(status.code()),
-        status.message().to_string(),
-        details,
-        metadata,
-    ))
-}
-
-/// Maps a `grpc-rust` [`StatusCodeError`] to the corresponding [`tonic::Code`].
-// TODO(#5991): Consider skipping conversion to `tonic::Code` by mapping
-// directly to `google_cloud_gax` error types
-// (https://github.com/googleapis/google-cloud-rust/pull/6082#discussion_r3603786027).
-fn grpc_rust_error_to_tonic_code(code: StatusCodeError) -> tonic::Code {
-    match code {
-        StatusCodeError::Cancelled => tonic::Code::Cancelled,
-        StatusCodeError::Unknown => tonic::Code::Unknown,
-        StatusCodeError::InvalidArgument => tonic::Code::InvalidArgument,
-        StatusCodeError::DeadlineExceeded => tonic::Code::DeadlineExceeded,
-        StatusCodeError::NotFound => tonic::Code::NotFound,
-        StatusCodeError::AlreadyExists => tonic::Code::AlreadyExists,
-        StatusCodeError::PermissionDenied => tonic::Code::PermissionDenied,
-        StatusCodeError::ResourceExhausted => tonic::Code::ResourceExhausted,
-        StatusCodeError::FailedPrecondition => tonic::Code::FailedPrecondition,
-        StatusCodeError::Aborted => tonic::Code::Aborted,
-        StatusCodeError::OutOfRange => tonic::Code::OutOfRange,
-        StatusCodeError::Unimplemented => tonic::Code::Unimplemented,
-        StatusCodeError::Internal => tonic::Code::Internal,
-        StatusCodeError::Unavailable => tonic::Code::Unavailable,
-        StatusCodeError::DataLoss => tonic::Code::DataLoss,
-        StatusCodeError::Unauthenticated => tonic::Code::Unauthenticated,
-    }
-}
-
-/// A background task handle that pumps responses from a `grpc-rust` [`RecvStream`].
-///
-/// This background pump is required because [`RecvStream::recv`](grpc::client::RecvStream::recv) is not cancellation-safe.
-/// If an outer caller cancels an async operation (e.g., via a timeout, `tokio::select!`, or dropping a stream early) that directly
-/// uses `RecvStream::recv`, the stream could become corrupted.
-///
-/// Isolating `RecvStream` in a dedicated background task ensures `recv` calls run to completion, while aborting the task
-/// on drop cleanly releases the owned [`RecvStream`]. Because `RecvStream` is owned by the background loop,
-/// `recv()` is never interrupted mid-call by outer task drops. Downstream application code only reads from a
-/// [`tokio::sync::mpsc::Receiver`], which is completely safe to drop or cancel at any time.
-pub struct ReceiveTask {
-    handle: tokio::task::JoinHandle<()>,
-}
-
-impl ReceiveTask {
-    /// Spawns a background task that pumps responses from `recv` into a bounded channel.
-    ///
-    /// Returns a channel receiver for responses and a [`ReceiveTask`] handle for managing task cancellation.
-    pub fn spawn<Response, R>(
-        recv: R,
-    ) -> (
-        tokio::sync::mpsc::Receiver<tonic::Result<Option<Response>>>,
-        Self,
-    )
-    where
-        Response: Message + Default + Send + 'static,
-        R: RecvStream + 'static,
-    {
-        let (tx, rx) = tokio::sync::mpsc::channel(1);
-        let handle = tokio::spawn(receive_responses(recv, tx));
-        (rx, Self { handle })
-    }
-}
-
-impl Drop for ReceiveTask {
-    fn drop(&mut self) {
-        // The task owns the RecvStream, so aborting drops that too.
-        self.handle.abort();
-    }
-}
-
-/// Drives the response pump loop
-async fn receive_responses<Response, R>(
-    mut recv: R,
-    tx: tokio::sync::mpsc::Sender<tonic::Result<Option<Response>>>,
-) where
-    Response: Message + Default + Send + 'static,
-    R: RecvStream + 'static,
-{
-    // A container into which incoming message payloads are decoded.
-    let mut slot = GrpcRustRecv::<Response>::default();
-    loop {
-        // Reserve before receiving to apply backpressure.
-        let Ok(permit) = tx.reserve().await else {
-            return;
-        };
-        let (response, is_terminal) = match recv.recv(&mut slot).await {
-            // TODO(#5991): Headers will be processed separately in a later PR.
-            ResponseStreamItem::Headers(_) => continue,
-            ResponseStreamItem::Message => match slot.take() {
-                Ok(message) => (Ok(Some(message)), false),
-                // A missing/undecodable message breaks the stream sequence
-                // and renders it unrecoverable.
-                Err(status) => (Err(status), true),
-            },
-            ResponseStreamItem::Trailers(trailers) => (
-                trailers_to_tonic_status(trailers).map_or(Ok(None), Err),
-                true,
-            ),
-            ResponseStreamItem::StreamClosed => (
-                Err(tonic::Status::internal(
-                    "grpc-rust response stream closed without trailers",
-                )),
-                true,
-            ),
-        };
-        permit.send(response);
-        if is_terminal {
-            return;
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use grpc::StatusError;
-    use grpc::metadata::MetadataValue;
     use pretty_assertions::assert_eq;
-    use test_case::test_case;
+    use prost::Message;
 
     #[derive(Clone, PartialEq, prost::Message)]
     struct TestMessage {
@@ -212,66 +51,10 @@ mod tests {
         let mut encoded = GrpcRustSend(want.clone())
             .encode()
             .map_err(anyhow::Error::msg)?;
-        let mut decoded = GrpcRustRecv::<TestMessage>::new();
-        decoded.decode(&mut encoded).map_err(anyhow::Error::msg)?;
-        let got = decoded.take()?;
+        let got = TestMessage::decode(encoded.copy_to_bytes(encoded.remaining()))?;
 
         // Assert
         assert_eq!(got, want);
         Ok(())
-    }
-
-    #[test]
-    fn trailers_to_tonic_status_preserves_details_and_metadata() {
-        // Arrange
-        const TEST_HEADER: &str = "x-test-header";
-        const TEST_VALUE: &str = "test-value";
-        const GRPC_STATUS_DETAILS_BIN: &str = "grpc-status-details-bin";
-        const ERROR_MESSAGE_DENIED: &str = "denied";
-
-        let details = b"status-details";
-        let mut metadata = grpc::metadata::MetadataMap::new();
-        metadata.insert_bin(GRPC_STATUS_DETAILS_BIN, MetadataValue::from_bytes(details));
-        metadata.insert(TEST_HEADER, MetadataValue::from_static(TEST_VALUE));
-        let trailers = Trailers::new(Err(StatusError::new(
-            StatusCodeError::PermissionDenied,
-            ERROR_MESSAGE_DENIED,
-        )))
-        .with_metadata(metadata);
-
-        // Act
-        let got = trailers_to_tonic_status(trailers).expect("non-OK trailers should map to status");
-
-        // Assert
-        assert_eq!(got.code(), tonic::Code::PermissionDenied);
-        assert_eq!(got.message(), ERROR_MESSAGE_DENIED);
-        assert_eq!(got.details(), details);
-        assert_eq!(
-            got.metadata()
-                .get(TEST_HEADER)
-                .and_then(|value| value.to_str().ok()),
-            Some(TEST_VALUE)
-        );
-    }
-
-    #[test_case(StatusCodeError::Cancelled, tonic::Code::Cancelled)]
-    #[test_case(StatusCodeError::Unknown, tonic::Code::Unknown)]
-    #[test_case(StatusCodeError::InvalidArgument, tonic::Code::InvalidArgument)]
-    #[test_case(StatusCodeError::DeadlineExceeded, tonic::Code::DeadlineExceeded)]
-    #[test_case(StatusCodeError::NotFound, tonic::Code::NotFound)]
-    #[test_case(StatusCodeError::AlreadyExists, tonic::Code::AlreadyExists)]
-    #[test_case(StatusCodeError::PermissionDenied, tonic::Code::PermissionDenied)]
-    #[test_case(StatusCodeError::ResourceExhausted, tonic::Code::ResourceExhausted)]
-    #[test_case(StatusCodeError::FailedPrecondition, tonic::Code::FailedPrecondition)]
-    #[test_case(StatusCodeError::Aborted, tonic::Code::Aborted)]
-    #[test_case(StatusCodeError::OutOfRange, tonic::Code::OutOfRange)]
-    #[test_case(StatusCodeError::Unimplemented, tonic::Code::Unimplemented)]
-    #[test_case(StatusCodeError::Internal, tonic::Code::Internal)]
-    #[test_case(StatusCodeError::Unavailable, tonic::Code::Unavailable)]
-    #[test_case(StatusCodeError::DataLoss, tonic::Code::DataLoss)]
-    #[test_case(StatusCodeError::Unauthenticated, tonic::Code::Unauthenticated)]
-    fn grpc_rust_error_to_tonic_code_maps_correctly(input: StatusCodeError, want: tonic::Code) {
-        // Act & Assert
-        assert_eq!(grpc_rust_error_to_tonic_code(input), want);
     }
 }
